Moving ui content around, and have dynamic handler to help ui development.
Note: idea is that in future we will use the Vue tool chain to build static pages into './ui/dist' - link this to cargo so that it's done before cargo starts compiling rust. And then include these files as static resources. Also, only expose /ui/dev in development mode (to be introduced, perhaps have an ENV for this, so we do not pollute the config).