Support publishing by Trust Anchor.
Collateral damage is that the next update time in repository stats
had to be removed. But, as it turned out that time was only used
in the API and metrics - not the UI - and it was inaccurate as
it did not take the threshold time before mft expiry into account.