Remove read lock when returning content.
This should not be needed. And it might lead to write lock starvation.
Typically cliens would call list before sending updates and not send
two requests at the same time. In case we don't have the read lock then
the content may not contain an update for another publisher, but that
is irrelevant here.