Rename timing_config to ta_config and alias (#1242)
This PR renames the `timing_config` options in the Krill config to
`ta_timing`. It was already required to have this name if the config
was used by the Krill TA signer. The `timing_config` name is allowed
as an alias in both cases now.

This a backport of #1241 to the 0.14 series of Krill.